+++ src/fastapi_framework_mvc/Server/logging.py
@@ -1,0 +1,24 @@
+# coding: utf-8
+
+
+def configure_logs(name, format, output_file, debug='info'):
+    """
+    Setup logging configuration, logging handler, logging formatting, ...
+    :param name:
+    :type name: str
+    :param format:
+    :type format: str
+    :param output_file:
+    :type output_file: str
+    :param debug:
+    :type debug: str
+    :return:
+    """
+    import logging
+    logger = logging.getLogger(name)
+    formatter = logging.Formatter(format)
+    file_handler = logging.FileHandler(output_file)
+    file_handler.setFormatter(formatter)
+    logger.addHandler(file_handler)
+    logging.getLevelName(debug.upper())
+    logger.setLevel(logging.getLevelName(debug.upper()))
